Code dump:
- Lots of changes to the Object system to simplify Python integration.
- Module management: Management of dynamically linked modules.
- Started integrating Flow: Editor, underlying graph structure, python
  integration, etc.
- A lot of small fixes, including console touch-ups.

+++ Source/Core/Object/Class.cpp
@@ -3,7 +3,7 @@
 #include "Class.h"
 
 
-Class::Class(const char* name, uint32_t size, CreateObjectFn creator)
+Class::Class(const char* name, size_t size, CreateObjectFn creator)
     : Type(name, size),
     _creator(creator),
     _super(nullptr)
@@ -31,4 +31,3 @@ Object* Class::create_object() const
 {
     return _creator();
 }
-
